# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/auth/create-employee": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Create Employee",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SA - Employee"
                ],
                "summary": "Create Employee",
                "parameters": [
                    {
                        "description": "CreateEmployeeRequest Request",
                        "name": "_",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.CreateEmployeeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiEmptyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    }
                }
            }
        },
...
        "/auth/list-employee": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "List Employee",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SA - Employee"
                ],
                "summary": "List Employee",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "name",
                            "code"
                        ],
                        "type": "string",
                        "description": "Sort field. Default to code,asc",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Ignore pagination and return all data",
                        "name": "paging_ignored",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by department id",
                        "name": "department_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by name or email, code, phone number",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by employee id",
                        "name": "employee_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by employee type 1: Doctor, 2: Nurse",
                        "name": "employee_type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/authmodel.ListEmployee"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    }
                }
            }
        },
...
        "/patient/medication/list-medication": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "List Medication",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SA - Medication Management"
                ],
                "summary": "List Medication",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "name",
                            "created_at"
                        ],
                        "type": "string",
                        "description": "Sort field. Default to name,asc",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Ignore pagination and return all data",
                        "name": "paging_ignored",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by name",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/patientmodel.ListMedicationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    }
                }
            }
        },
        "/patient/medication/upsert-medication": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Upsert Medication",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SA - Medication Management"
                ],
                "summary": "Upsert Medication",
                "parameters": [
                    {
                        "description": "Upsert medication Request",
                        "name": "_",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/patient.UpsertMedicationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiEmptyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "auth.CreateEmployeeRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "date_of_birth": {
                    "type": "integer"
                },
                "degree_name": {
                    "type": "string"
                },
                "degree_year": {
                    "type": "integer"
                },
                "department_id": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "start_date": {
                    "type": "integer"
                },
                "type": {
                    "type": "integer"
                }
            }
        },
        "auth.Department": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "auth.User": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "date_of_birth": {
                    "type": "integer"
                },
                "degree_name": {
                    "type": "string"
                },
                "degree_year": {
                    "type": "integer"
                },
                "department": {
                    "$ref": "#/definitions/auth.Department"
                },
                "end_date": {
                    "description": "Off board date",
                    "type": "integer"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "start_date": {
                    "description": "Onboard date",
                    "type": "integer"
                },
                "state": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "authmodel.Employees": {
            "type": "object",
            "properties": {
                "employees": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/auth.User"
                    }
                },
                "hasNext": {
                    "type": "boolean"
                },
                "hasPrevious": {
                    "type": "boolean"
                },
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "totalElements": {
                    "description": "Num items in all pages",
                    "type": "integer"
                },
                "totalItems": {
                    "description": "Num items in current page",
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
...
        "authmodel.ListEmployee": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {
                    "$ref": "#/definitions/authmodel.Employees"
                }
            }
        },
...
        "basemodel.ApiEmptyResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "basemodel.ApiError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gcommon.ErrorDetail"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "gcommon.ErrorDetail": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "patient.Medication": {
            "type": "object",
            "properties": {
                "effects": {
                    "type": "string"
                },
                "expired_date": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
...
        "patient.UpsertMedicationRequest": {
            "type": "object",
            "properties": {
                "effects": {
                    "type": "string"
                },
                "expired_date": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
...
        "patientmodel.ListMedicationResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {
                    "$ref": "#/definitions/patientmodel.ListMedicationResponseData"
                }
            }
        },
        "patientmodel.ListMedicationResponseData": {
            "type": "object",
            "properties": {
                "hasNext": {
                    "type": "boolean"
                },
                "hasPrevious": {
                    "type": "boolean"
                },
                "medications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/patient.Medication"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "totalElements": {
                    "description": "Num items in all pages",
                    "type": "integer"
                },
                "totalItems": {
                    "description": "Num items in current page",
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
...
    }
}